+++ src/main/java/de/hhbk/jahresprojekt/model/builder/ItemBuilder.java
@@ -2,8 +2,12 @@ package de.hhbk.jahresprojekt.model.builder;
 
 import de.hhbk.jahresprojekt.model.Item;
 
+/**
+ * @author Frederik Hafemann
+ * @author Enrico Messall
+ */
 public final class ItemBuilder {
-    private int id;
+    private Long id;
     private int value;
     private String comment;
 
@@ -14,7 +18,7 @@ public final class ItemBuilder {
         return new ItemBuilder();
     }
 
-    public ItemBuilder withId(int id) {
+    public ItemBuilder withId(Long id) {
         this.id = id;
         return this;
     }
